# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BountyPilot AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BountyStatus = typeof BountyStatus[keyof typeof BountyStatus];


export const BountyStatus = {
  discovered: 'discovered',
  saved_for_later: 'saved_for_later',
  approved: 'approved',
  rejected: 'rejected',
  researching: 'researching',
  scripting: 'scripting',
  recording: 'recording',
  editing: 'editing',
  submitted: 'submitted',
  won: 'won',
  lost: 'lost',
} as const;

export interface Bounty {
  id: number;
  url: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  platform?: string | null;
  /** @nullable */
  projectName?: string | null;
  /** @nullable */
  rewardAmount?: string | null;
  /** @nullable */
  rewardCurrency?: string | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  contentFormat?: string | null;
  /** @nullable */
  submissionRequirements?: string | null;
  /** @nullable */
  deliverables?: string | null;
  /** @nullable */
  submissionLink?: string | null;
  /** @nullable */
  eligibilityRules?: string | null;
  /** @nullable */
  importantNotes?: string | null;
  /** @nullable */
  opportunityScore?: number | null;
  /** @nullable */
  scoreExplanation?: string | null;
  status: BountyStatus;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface BountyInput {
  url: string;
}

export interface BountyUpdate {
  status?: string;
  title?: string;
  rewardAmount?: string;
  deadline?: string;
}

export interface ResearchBrief {
  id: number;
  bountyId: number;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  contentAngles?: string | null;
  /** @nullable */
  keyPoints?: string | null;
  /** @nullable */
  targetAudience?: string | null;
  /** @nullable */
  competitorAnalysis?: string | null;
  createdAt: string;
}

export interface ProductionPlan {
  id: number;
  bountyId: number;
  /** @nullable */
  scriptOutline?: string | null;
  /** @nullable */
  shotList?: string | null;
  /** @nullable */
  captionDraft?: string | null;
  /** @nullable */
  submissionChecklist?: string | null;
  /** @nullable */
  estimatedHours?: number | null;
  createdAt: string;
}

/**
 * @nullable
 */
export type SubmissionResult = typeof SubmissionResult[keyof typeof SubmissionResult] | null;


export const SubmissionResult = {
  pending: 'pending',
  won: 'won',
  lost: 'lost',
} as const;

export interface Submission {
  id: number;
  bountyId: number;
  /** @nullable */
  submittedAt?: string | null;
  /** @nullable */
  submissionUrl?: string | null;
  /** @nullable */
  result?: SubmissionResult;
  /** @nullable */
  rewardReceived?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface SubmissionInput {
  bountyId: number;
  submittedAt?: string;
  submissionUrl?: string;
  notes?: string;
}

export interface SubmissionUpdate {
  result?: string;
  rewardReceived?: number;
  notes?: string;
  submissionUrl?: string;
}

export interface Earning {
  id: number;
  /** @nullable */
  bountyId?: number | null;
  /** @nullable */
  platform?: string | null;
  amount?: number;
  currency?: string;
  /** @nullable */
  receivedAt?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface EarningInput {
  bountyId?: number;
  platform?: string;
  amount: number;
  currency: string;
  receivedAt?: string;
  notes?: string;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface DashboardSummary {
  totalBounties: number;
  totalEarnings: number;
  activeBounties: number;
  wonBounties: number;
  pipelineValue: number;
  /** @nullable */
  averageScore?: number | null;
  statusBreakdown?: StatusCount[];
}

export interface PlatformStat {
  platform: string;
  count: number;
  /** @nullable */
  totalReward?: number | null;
}

export type ListBountiesParams = {
status?: string;
platform?: string;
};