# lib/api-zod/src/generated/types/bounty.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BountyPilot AI API
 * OpenAPI spec version: 0.1.0
 */
import type { BountyStatus } from './bountyStatus';

export interface Bounty {
  id: number;
  url: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  platform?: string | null;
  /** @nullable */
  projectName?: string | null;
  /** @nullable */
  rewardAmount?: string | null;
  /** @nullable */
  rewardCurrency?: string | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  contentFormat?: string | null;
  /** @nullable */
  submissionRequirements?: string | null;
  /** @nullable */
  deliverables?: string | null;
  /** @nullable */
  submissionLink?: string | null;
  /** @nullable */
  eligibilityRules?: string | null;
  /** @nullable */
  importantNotes?: string | null;
  /** @nullable */
  opportunityScore?: number | null;
  /** @nullable */
  scoreExplanation?: string | null;
  status: BountyStatus;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}