# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BountyPilot AI API
 * OpenAPI spec version: 0.1.0
 */
import type { StatusCount } from './statusCount';

export interface DashboardSummary {
  totalBounties: number;
  totalEarnings: number;
  activeBounties: number;
  wonBounties: number;
  pipelineValue: number;
  /** @nullable */
  averageScore?: number | null;
  statusBreakdown?: StatusCount[];
}