# lib/api-zod/src/generated/types/earningInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BountyPilot AI API
 * OpenAPI spec version: 0.1.0
 */

export interface EarningInput {
  bountyId?: number;
  platform?: string;
  amount: number;
  currency: string;
  receivedAt?: string;
  notes?: string;
}