# lib/api-zod/src/generated/types/platformStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BountyPilot AI API
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformStat {
  platform: string;
  count: number;
  /** @nullable */
  totalReward?: number | null;
}