# lib/api-zod/src/generated/types/submission.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BountyPilot AI API
 * OpenAPI spec version: 0.1.0
 */
import type { SubmissionResult } from './submissionResult';

export interface Submission {
  id: number;
  bountyId: number;
  /** @nullable */
  submittedAt?: string | null;
  /** @nullable */
  submissionUrl?: string | null;
  /** @nullable */
  result?: SubmissionResult;
  /** @nullable */
  rewardReceived?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}