# lib/api-zod/src/generated/types/submissionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BountyPilot AI API
 * OpenAPI spec version: 0.1.0
 */

export interface SubmissionInput {
  bountyId: number;
  submittedAt?: string;
  submissionUrl?: string;
  notes?: string;
}